--- a/Makefile
+++ b/Makefile
@@ -21,6 +21,6 @@ check-go:
 		echo go fmt is sad: $(GOFMT); \
 		exit 1; \
 	fi )
-	@(go tool vet -all -composites=false -copylocks=false .)
+	@(go vet -all .)
 
 .PHONY: default check docs check-licence check-go